Added Pumpkin and Moon Melon Crop, Commit 2

+++ code/level.py
@@ -129,8 +129,9 @@ class Level:
 
 		# apples on the trees
 		for tree in self.tree_sprites.sprites():
-			for apple in tree.apple_sprites.sprites():
-				apple.kill()
+			if hasattr(tree, 'apple_sprites'):
+				for apple in tree.apple_sprites.sprites():
+					apple.kill()
 			tree.create_fruit()
 
 		# sky
